Move signUp function to registrationActions
