+++ client/src/actions/sessionActions.ts
@@ -5,7 +5,7 @@ import * as H from 'history'
 import { setting as axios } from 'config/axios'
 import * as actionTypes from 'utils/actionTypes'
 import { setCookies, clearCookies } from 'utils/cookies'
-import { User, LoginParams, SignUpParams } from 'types/api'
+import { User, LoginParams } from 'types/api'
 import { CookiesHeader } from 'types/store'
 import { UserAction } from 'types/action'
 
@@ -50,36 +50,3 @@ export const logout = (): Action => {
   }
 }
 
-export const signUpRequest = (): Action => {
-  return {
-    type: actionTypes.SIGN_UP_REQUEST
-  }
-}
-
-export const signUpSuccess = (user: User, headers: CookiesHeader): UserAction => {
-  setCookies(headers)
-  return {
-    type: actionTypes.SIGN_UP_SUCCESS,
-    user
-  }
-}
-
-export const signUpFailure = (): Action => {
-  return {
-    type: actionTypes.SIGN_UP_FAILURE
-  }
-}
-
-export const signUp = (params: SignUpParams) => {
-  return async (dispatch: Dispatch<Action>): Promise<void> => {
-    dispatch(signUpRequest())
-    try {
-      const res = await axios.post('/api/auth', params)
-      console.log(res)
-      return dispatch(signUpSuccess(res.data, res.headers))
-    }
-    catch(err) {
-      return dispatch(signUpFailure())
-    }
-  }
-}
